requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
cassandra-driver>=3.29.1
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
slowapi>=0.1.9
prometheus-fastapi-instrumentator>=7.0.0
fastapi-mail>=1.4.1
pydantic>=2.0.0
pydantic-settings>=2.0.0